+++ src/ft_strmap.c
@@ -6,7 +6,7 @@
 /*   By: rhallste <rhallste@student.42.us.org>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/19 17:00:22 by rhallste          #+#    #+#             */
-/*   Updated: 2017/09/23 23:02:46 by rhallste         ###   ########.fr       */
+/*   Updated: 2017/11/22 10:13:10 by rhallste         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 char	*ft_strmap(char const *s, char (*f)(char))
 {
-	int		len;
-	int		i;
+	size_t	len;
+	size_t	i;
 	char	*new;
 
 	if (s)
